Added operator overloading for Decimals Note: This feature is being enabled through babel, and unfortunately doesn't really have any typescript support. Using an overloaded operator will show an error and be typed as "any". If ecmascript ever support operator overloading, then typescript will follow suit and these issues can be resolved. Here's the current proposal for how that could look like, although it's a long way's off from being accepted, if it ever is: https://github.com/tc39/proposal-operator-overloading
Alternatively, there's a proposal for declaring that certain types have
operator overloads, which would also work just perfectly:
https://github.com/microsoft/TypeScript/issues/42218

In the meantime, the errors will unfortunately remain present, although
they won't cause any issues in production.

BTW, the rhs can be any DecimalSource, but the lhs has to be a Decimal.